fix: divide simulation config on different files
