+++ services/environment-api/src/config_schema.py
@@ -4,18 +4,14 @@ from typing import Literal
 
 from common.utils.settings_base import BaseSettings
 from pydantic import BaseModel, ConfigDict, Field
-
-from .config_simulation_schema import (
-    CatalogConfig,
-    InfrastructureConfig,
-    LayoutConfig,
-    LogisticsConfig,
-    OutboundConfig,
-    ReplenishmentConfig,
-    SensorsConfig,
-    SimulationConfig,
-    WorldConfig,
-)
+from src.simulation_configs.catalog import CatalogConfig
+from src.simulation_configs.infrastructure import InfrastructureConfig
+from src.simulation_configs.layout import LayoutConfig
+from src.simulation_configs.logistics import LogisticsConfig
+from src.simulation_configs.outbound import OutboundConfig
+from src.simulation_configs.replenishment import ReplenishmentConfig
+from src.simulation_configs.sensors import SensorsConfig
+from src.simulation_configs.world import SimulationConfig, WorldConfig
 
 
 class AppConfig(BaseModel):
@@ -47,6 +43,7 @@ class Settings(BaseSettings):
     app: AppConfig = Field(default_factory=AppConfig)
     server: ServerConfig = Field(default_factory=ServerConfig)
     logging: LoggingConfig = Field(default_factory=LoggingConfig)
+
     world: WorldConfig
     simulation: SimulationConfig
     catalog: CatalogConfig
