+++ services/environment-api/src/simulation_configs/infrastructure.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel, ConfigDict
+
+
+class InfrastructureConfig(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+
+    region_timezones: dict[str, str]
